requirements.txt
langchain_google_genai
langchain_community
crawl4ai
playwright
streamlit
